+++ src/native-common/AccessibilityUtil.ts
@@ -15,14 +15,14 @@ import { AccessibilityUtil as CommonAccessibilityUtil, AccessibilityPlatformUtil
 
 import Types = require('../common/Types');
 
-const liveRegionMap = {
+const liveRegionMap: { [key: string]: string } = {
     [Types.AccessibilityLiveRegion.None]: 'none',
     [Types.AccessibilityLiveRegion.Assertive]: 'assertive',
     [Types.AccessibilityLiveRegion.Polite]: 'polite'
 };
 
 // iOS supported map.
-const traitsMap = {
+const traitsMap: { [key: string]: string } = {
     [Types.AccessibilityTrait.None]: 'none',
     [Types.AccessibilityTrait.Tab]: 'none', // NOTE: Tab trait isn't supported on iOS. Setting it to none, allows us to give it a custom
                                             // label. This needs to be done for any custom role, which needs to be supported on iOS.
@@ -45,7 +45,7 @@ const traitsMap = {
 };
 
 // Android supported map.
-const componentTypeMap = {
+const componentTypeMap: { [key: string]: string } = {
     [Types.AccessibilityTrait.None]: 'none',
     [Types.AccessibilityTrait.Tab]: 'none', // NOTE: Tab component type isn't supported on Android. Setting it to none, allows us to give
                                             // it a custom label. This needs to be done for any custom role, which needs to be supported