+++ src/web/AccessibilityUtil.ts
@@ -15,7 +15,7 @@ import Types = require('../common/Types');
 
 // Map of accessibility trait to an aria role attribute.  
 // What's a role attribute? https://www.w3.org/wiki/PF/XTech/HTML5/RoleAttribute
-const roleMap = {
+const roleMap: { [key: string]: string } = {
     [Types.AccessibilityTrait.None]: 'presentation',
     [Types.AccessibilityTrait.Button]: 'button',
     [Types.AccessibilityTrait.Link]: 'link',
@@ -42,7 +42,7 @@ const roleMap = {
 }; 
 
 // Map of accesssibility live region to an aria-live property.
-const liveRegionMap = {
+const liveRegionMap: { [key: string]: string } = {
     [Types.AccessibilityLiveRegion.None]: 'off',
     [Types.AccessibilityLiveRegion.Assertive]: 'assertive',
     [Types.AccessibilityLiveRegion.Polite]: 'polite'